+++ src/Sentry.Profiling/SampleProfilerSession.cs
@@ -12,6 +12,7 @@ namespace Sentry.Profiling;
 internal class SampleProfilerSession : IDisposable
 {
     private readonly EventPipeSession _session;
+    private readonly TraceLogEventSource _eventSource;
     private readonly SampleProfilerTraceEventParser _sampleEventParser;
     private readonly IDiagnosticLogger? _logger;
     private readonly SentryStopwatch _stopwatch;
@@ -22,8 +23,8 @@ internal class SampleProfilerSession : IDisposable
     {
         _session = session;
         _logger = logger;
-        EventSource = eventSource;
-        _sampleEventParser = new SampleProfilerTraceEventParser(EventSource);
+        _eventSource = eventSource;
+        _sampleEventParser = new SampleProfilerTraceEventParser(_eventSource);
         _stopwatch = stopwatch;
         _processing = processing;
     }
@@ -31,13 +32,13 @@ internal class SampleProfilerSession : IDisposable
     // Exposed only for benchmarks.
     internal static EventPipeProvider[] Providers = new[]
     {
-        new EventPipeProvider(ClrTraceEventParser.ProviderName, EventLevel.Verbose, (long) (
-            ClrTraceEventParser.Keywords.Jit
-            | ClrTraceEventParser.Keywords.NGen
-            | ClrTraceEventParser.Keywords.Loader
-            | ClrTraceEventParser.Keywords.Binder
-            | ClrTraceEventParser.Keywords.JittedMethodILToNativeMap
-            )),
+        // Note: all events we need issued by "DotNETRuntime" provider are at "EventLevel.Informational"
+        // see https://learn.microsoft.com/en-us/dotnet/fundamentals/diagnostics/runtime-events
+        // TODO replace Keywords.Default with a subset. Currently it is:
+        //   Default = GC | Type | GCHeapSurvivalAndMovement | Binder | Loader | Jit | NGen | SupressNGen
+        //                | StopEnumeration | Security | AppDomainResourceManagement | Exception | Threading | Contention | Stack | JittedMethodILToNativeMap
+        //                | ThreadTransfer | GCHeapAndTypeNames | Codesymbols | Compilation,
+        new EventPipeProvider(ClrTraceEventParser.ProviderName, EventLevel.Informational, (long) ClrTraceEventParser.Keywords.Default),
         new EventPipeProvider(SampleProfilerTraceEventParser.ProviderName, EventLevel.Informational),
         // new EventPipeProvider(TplEtwProviderTraceEventParser.ProviderName, EventLevel.Informational, (long) TplEtwProviderTraceEventParser.Keywords.Default)
     };
@@ -47,14 +48,11 @@ internal class SampleProfilerSession : IDisposable
     // need a large buffer if we're connecting righ away. Leaving it too large increases app memory usage.
     internal static int CircularBufferMB = 16;
 
-    // Exposed for tests
-    internal TraceLogEventSource EventSource { get; }
-
     public SampleProfilerTraceEventParser SampleEventParser => _sampleEventParser;
 
     public TimeSpan Elapsed => _stopwatch.Elapsed;
 
-    public TraceLog TraceLog => EventSource.TraceLog;
+    public TraceLog TraceLog => _eventSource.TraceLog;
 
     // default is false, set 1 for true.
     private static int _throwOnNextStartupForTests = 0;
@@ -112,7 +110,7 @@ internal class SampleProfilerSession : IDisposable
     {
         var tcs = new TaskCompletionSource();
         var cb = (TraceEvent _) => { tcs.TrySetResult(); };
-        EventSource.AllEvents += cb;
+        _eventSource.AllEvents += cb;
         try
         {
             // Wait for the first event to be processed.
@@ -120,7 +118,7 @@ internal class SampleProfilerSession : IDisposable
         }
         finally
         {
-            EventSource.AllEvents -= cb;
+            _eventSource.AllEvents -= cb;
         }
     }
 
@@ -134,7 +132,7 @@ internal class SampleProfilerSession : IDisposable
                 _session.Stop();
                 _processing.Wait();
                 _session.Dispose();
-                EventSource.Dispose();
+                _eventSource.Dispose();
             }
             catch (Exception ex)
             {