+++ src/Sentry/Platforms/Cocoa/CocoaScopeObserver.cs
@@ -67,6 +67,12 @@ internal sealed class CocoaScopeObserver : IScopeObserver
 
     public void SetTag(string key, string value)
     {
+        if (value is null)
+        {
+            _options.LogDebug("Tag with key '{0}' was null. Use Unset to remove tags instead.", key);
+            return;
+        }
+
         try
         {
             SentryCocoaSdk.ConfigureScope(scope => scope.SetTagValue(value, key));